Repository for RogerNav and RogerCMD Subsystems
Kayak Navigation System Description, Design, Development, and Testing
related files and information.

KayakNav subsystems are named in honor of Roger Keeney (founder of
A.I.R.S., a 501(c)3 non-profit), who initially conceived of the idea for
the development of a system to enable the visually disabled to compete
in kayak racing, leading to the development, in 2013, of the "Course
Keeper" by volunteers of the Atlanta HACEMOS Chapter, another 501c(3).

This system is being developed by Atlanta HACEMOS Chapter volunteers to
assist Ahmet Ustunel, a 2017 Holman Prize recipient, with his goal to be
the first blind person to kayak from Europe to Asia (or vice versa)
solo.